refactor: mv datanode configs to a datanode-options crate
