+++ src/datanode-options/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "datanode-options"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+common-base = { workspace = true }
+common-procedure = { workspace = true }
+common-telemetry = { workspace = true }
+humantime-serde = "1.1"
+meta-client = { workspace = true }
+secrecy = { version = "0.8", features = ["serde", "alloc"] }
+serde.workspace = true
+servers = { workspace = true }
+storage = { workspace = true }
+toml.workspace = true
