+++ src/datanode/src/server.rs
@@ -17,6 +17,7 @@ use std::net::SocketAddr;
 use std::sync::Arc;
 
 use common_runtime::Builder as RuntimeBuilder;
+use datanode_options::DatanodeOptions;
 use futures::future;
 use servers::grpc::GrpcServer;
 use servers::http::{HttpServer, HttpServerBuilder};
@@ -25,7 +26,6 @@ use servers::query_handler::grpc::ServerGrpcQueryHandlerAdaptor;
 use servers::server::Server;
 use snafu::ResultExt;
 
-use crate::datanode::DatanodeOptions;
 use crate::error::{
     ParseAddrSnafu, Result, RuntimeResourceSnafu, ShutdownServerSnafu, StartServerSnafu,
     WaitForGrpcServingSnafu,
