+++ src/datanode/src/store.rs
@@ -25,12 +25,12 @@ use std::sync::Arc;
 
 use common_base::readable_size::ReadableSize;
 use common_telemetry::logging::info;
+use datanode_options::{ObjectStoreConfig, DEFAULT_OBJECT_STORE_CACHE_SIZE};
 use object_store::layers::{LoggingLayer, LruCacheLayer, MetricsLayer, RetryLayer, TracingLayer};
 use object_store::services::Fs as FsBuilder;
 use object_store::{util, ObjectStore, ObjectStoreBuilder};
 use snafu::prelude::*;
 
-use crate::datanode::{ObjectStoreConfig, DEFAULT_OBJECT_STORE_CACHE_SIZE};
 use crate::error::{self, Result};
 
 pub(crate) async fn new_object_store(store_config: &ObjectStoreConfig) -> Result<ObjectStore> {
