+++ src/datanode/src/store/fs.rs
@@ -15,11 +15,11 @@
 use std::{fs, path};
 
 use common_telemetry::logging::info;
+use datanode_options::FileConfig;
 use object_store::services::Fs as FsBuilder;
 use object_store::{util, ObjectStore};
 use snafu::prelude::*;
 
-use crate::datanode::FileConfig;
 use crate::error::{self, Result};
 use crate::store;
