+++ src/datanode/src/store/gcs.rs
@@ -13,12 +13,12 @@
 // limitations under the License.
 
 use common_telemetry::logging::info;
+use datanode_options::GcsConfig;
 use object_store::services::Gcs as GCSBuilder;
 use object_store::{util, ObjectStore};
 use secrecy::ExposeSecret;
 use snafu::prelude::*;
 
-use crate::datanode::GcsConfig;
 use crate::error::{self, Result};
 
 pub(crate) async fn new_gcs_object_store(gcs_config: &GcsConfig) -> Result<ObjectStore> {
