+++ src/datanode/src/store/s3.rs
@@ -13,12 +13,12 @@
 // limitations under the License.
 
 use common_telemetry::logging::info;
+use datanode_options::S3Config;
 use object_store::services::S3 as S3Builder;
 use object_store::{util, ObjectStore};
 use secrecy::ExposeSecret;
 use snafu::prelude::*;
 
-use crate::datanode::S3Config;
 use crate::error::{self, Result};
 
 pub(crate) async fn new_s3_object_store(s3_config: &S3Config) -> Result<ObjectStore> {
